Currency becomes a site-wide admin setting, not a per-user one
The picker moves to Admin > Settings and now drives every price on the site —
billing plans, the public /pricing page, and the landing pricing section. The
user-facing Settings page, its route, nav entry and users.display_currency are
gone.

WHY THE DISCLOSURE GOT LOUDER, NOT QUIETER. A per-user preference was chosen
by the person reading it, next to a warning they had just seen. A site-wide
one applies to visitors who never chose anything — so the site can show
"$51.79" to someone who will be charged in rupees and has no idea. Both
pricing surfaces now print the ₹ figure whenever the displayed price is a
conversion ("approx. — billed ₹4,999"), and the admin card says plainly that
this changes display only.

PUBLIC endpoint, deliberately. /pricing and the landing page are what most
visitors see; an authed-only source would have left them on the base currency
while the app followed the admin's choice — the same product priced two ways
depending on whether you were logged in.

Validation moved with the setting. The admin update applies fields with a
generic setattr loop, so without a validator an unsupported code would save,
report success, and leave every price silently on INR — a setting that looks
applied and isn't. SUPPORTED_CURRENCIES now lives in services/fx.py, which
both the schema validator and the public endpoint read.

The JSON-LD offers and payment history stay INR on purpose: one describes the
actual offer, the other records completed charges. Verified in the built
output — priceCurrency is INR and nothing else.

Prerendered pages render ₹ because the build has no API to ask; the client
converts on load. That is the right fallback, and confirmed in dist/.

Verified against a running server, and worth recording HOW: the first check
404'd on the new routes. The cause was not the code — an older uvicorn from
earlier in the session still held port 8000, so my new process failed to bind
("only one usage of each socket address") and the stale one answered. Retried
on a free port: routes registered, and flipping the DB value gave USD 0.01036,
JPY 1.684463, EUR 0.009078, and unset -> INR with converted:false.

58 frontend tests, 206 backend, lint and build clean.

+++ backend/app/schemas/ai_visibility.py
@@ -41,7 +41,7 @@ class AiVisibilitySummary(BaseModel):
     ai_overview_present: int = 0   # keywords that triggered an AI Overview
     ai_overview_cited: int = 0     # ...where your domain is cited
     ai_mode_cited: int = 0         # keywords where you're cited in AI Mode
-    cost_cents: int = 0
+    cost_cents: float = 0
 
 
 class AiVisibilityStatusResponse(BaseModel):
